共通Fetcherを用いてAPIにアクセスするように修正
